+++ content/src/main/content/jcr_root/apps/acs-commons/authoring/shared-component-properties/touchui/js/shared-component-properties.js
@@ -116,5 +116,15 @@
             ns.EditorFrame.editableToolbar.registerAction('GLOBAL-COMPONENT-PROPS', initDialog("global"));
         }
     });
+    // AEM 6.3 Bug Fix for #1982
+    // we listen to the editables to reset the Shared Component Properties Registered flag so that the Shared and
+    // Global properties are loaded as page refreshes due to Edit Config Listeners
+    channel.on('cq-editables-updated', function (ev) {
+        // reset the flag so that the shared and global icons are loaded again after page refresh
+        if (ns.acsSharedComponentPropertiesIsRegistered &&
+            ns.EditorFrame && ns.EditorFrame.editableToolbar && ns.EditorFrame.editableToolbar._customActions) {
+            ns.acsSharedComponentPropertiesIsRegistered = false;
+        }
+    });
 
 }(jQuery, Granite.author, jQuery(document), this));