+++ CHANGELOG.md
@@ -17,7 +17,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 - #1945 - Added support for jcr:content creation and update to the Data Importer
 
 ### Fixed
+- #1547 - Updated Report Runner's ReportCSVExportServlet to support extra-ACS Commons ReportCellCSVExporter
 - #1976 - Fixed failing Remote Assets and tests dependent on mock server on JDK 11
+- #1982 - Fixed the Shared and Global icons that are not appearing in edit bar when the dialog is edited and saved and page refreshes due to Edit Config Listener ( Shared Component Properties )
 
 ## [4.2.0] - 2019-06-18